feat(templateApi): 템플릿 데이터 가져오는 api 작성
